feat(interface-schema): separate mongo based schemas and actual interface
