+++ src/lib/api/auth.ts
@@ -1,19 +1,15 @@
 import axios from "lib/axios";
 
-import type { User, GroupApply } from "lib/interface";
-
-interface UserWithPendingGroups extends User {
-  pendingGroups?: Pick<GroupApply, "_id" | "name" | "profilePhoto" | "subtitle">[];
-}
+import type { User } from "lib/interface";
 
 export const loginCallback = (code: string, state: string, update?: boolean) =>
   axios.post<{
     token: string;
-    user: UserWithPendingGroups;
+    user: User;
   }>("/auth/login/callback", {
     code,
     state,
     update,
   });
 
-export const checkAuth = () => axios.get<UserWithPendingGroups>("/auth");
+export const checkAuth = () => axios.get<User>("/auth");
