+++ src/lib/interface/group.ts
@@ -1,31 +1,7 @@
-import type { MongoSchema } from "./utils/mongo";
-import type { User } from ".";
+import type { Group as GroupSchema } from "./schemas";
 
-export interface GroupApply extends MongoSchema {
-  name: string;
-  subtitle?: string;
-  description?: string;
-  profilePhoto?: string;
-  backgroundPhoto?: string;
-  category?: string[];
-  members: {
-    user: User;
-    role?: "admin" | "editor";
-  }[];
-  purpose?: string;
-  isBusiness?: boolean;
-}
-
-export interface Group extends GroupApply {
-  isPreRegistered?: boolean;
-  level?: number;
-  revisionHistory?: RevisionHistory[];
-  recentUpload?: string;
-  followers?: User[] | string[];
-  score: number;
-}
-
-interface RevisionHistory extends MongoSchema {
-  prev?: string;
-  next?: string;
+export interface Group extends GroupSchema {
+  zabosCount?: number;
+  followersCount?: number;
+  isPending?: boolean;
 }
