+++ src/lib/interface/schemas/index.ts
@@ -1,2 +1,4 @@
+export * from "./board";
 export * from "./group";
 export * from "./user";
+export * from "./zabo";
